// src/charting_library.d.ts
export interface IContext {
    /**
     * Symbol Instrument
     */
    symbol: ISymbolInstrument;
    /**
     * Load a new symbol for the custom indicator
     * @param  {string} tickerid - Symbol identifier
     * @param  {string} period - period for the new symbol
     * @param  {string} currencyCode? - Currency code
     * @param  {string} unitId? - Unit id
     */
    new_sym(tickerid: string, period: string, currencyCode?: string, unitId?: string): ISymbolInstrument;
    /**
     * Switch context to the other symbol received through {@link IContext.new_sym}
     * @param  {number} i - the index of the symbol (`0` for the main series)
     */
    select_sym(i: number): void;
    /**
     * Creates an in-memory temporary storage with depth defined by the first call `new_var(value).get(n)`
     * @param  {number} value? - variable's value
     */
    new_var(value?: number): IPineSeries;
    /**
     * Creates an in-memory temporary storage with unlimited depth.
     * @param  {number} value? - variable's value
     */
    new_unlimited_var(value?: number): IPineSeries;
    /**
     * Creates a new context
     */
    new_ctx(): IContext;
    /**
     * Checks if symbol is the main symbol
     * @param  {ISymbolInstrument|undefined} symbol - symbol to check
     * @returns `true` if symbol is the main symbol
     */
    is_main_symbol(symbol: ISymbolInstrument | undefined): boolean;
}
// ...
export interface IPineSeries {
    /**
     * Get the value at a specific index.
     *
     * Note: The indices of a pine series is opposite.
     *
     * Example:
     * - s.get(1) returns second last,
     * - s.get(2) - third last
     * - and so on
     * @param  {number} n? - index
     */
    get(n?: number): number;
    /**
     * Set the value for the pine series at the current index interation.
     * @param  {number} value - value to be set
     */
    set(value: number): void;
    /**
     * Get the index for the bar at the specified timestamp
     * @param  {number} time - timestamp
     */
    indexOf(time: number): number;
}
// ...
export interface ISymbolInstrument {
    /** Period Base */
    periodBase: string;
    /** Ticker ID */
    tickerid: string;
    /** Currency Code */
    currencyCode?: string;
    /** Period */
    period: string;
    /** Index */
    index: number;
    /** Time */
    time: number;
    /** Open bar value */
    open: number;
    /** High bar value */
    high: number;
    /** Low bar value */
    low: number;
    /** Close bar value */
    close: number;
    /** Bar Volume value */
    volume: number;
    /** Time of the update */
    updatetime: number;
    /** Session string */
    session: string;
    /** Script */
    script: any; // tslint:disable-line:no-any
    /** Ticker */
    ticker: string;
    /** Resolution */
    resolution: string;
    /** Interval */
    interval: number;
    /** Minimum tick amount */
    minTick: number;
    /** Whether this is the first bar */
    isFirstBar: boolean;
    /** Whether this is the last bar */
    isLastBar: boolean;
    /** Whether this is a new bar */
    isNewBar: boolean;
    /** Whether the bar is closed */
    isBarClosed: boolean;
    /** Symbol information */
    info?: LibrarySymbolInfo;
}
// ...
export interface PineJS {
    Std: PineJSStd;
}